requirements.txt
fastapi
uvicorn
langchain
langchain-google-vertexai
langchain-google-genai
requests
python-dotenv
inference-sdk
python-multipart
python-jose
google-genai
